packages/ml-service/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leaf-disease-detection-ml"
version = "1.0.0"
description = "ML service for leaf disease detection using U-Net"
authors = [{name = "Leaf Disease Detection Team"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.103.1",
    "uvicorn[standard]==0.23.2",
    "python-multipart==0.0.6",
    "pydantic==2.3.0",
    "pydantic-settings==2.0.3",
    "tensorflow==2.13.0",
    "keras==2.13.1",
    "numpy==1.24.3",
    "opencv-python==4.8.0.76",
    "pillow==10.0.0",
    "scikit-learn==1.3.0",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1"
]

[project.optional-dependencies]
test = [
    "pytest==7.4.2",
    "pytest-asyncio==0.21.1",
    "hypothesis==6.87.0",
    "httpx==0.24.1"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]